removed colons from summary log messages, other fixes

+++ stop_ssds.sh
@@ -2,7 +2,13 @@
 
 # Stop the firewall daemons
 
-kill `cat run/http_daemon.pid`
-kill `cat run/sshd_daemon.pid`
-kill `cat run/mail_daemon.pid`
+if [ -f "run/http_daemon.pid" ] ; then
+   kill `cat run/http_daemon.pid`
+fi
+if [ -f "run/sshd_daemon.pid" ] ; then
+   kill `cat run/sshd_daemon.pid`
+fi
+if [ -f "run/mail_daemon.pid" ] ; then
+   kill `cat run/mail_daemon.pid`
+fi
